# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ChatCitation": {
        "description": "A numbered source the assistant's answer is grounded in.",
        "properties": {
          "index": {
            "title": "Index",
            "type": "integer"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "locator": {
            "default": "",
            "title": "Locator",
            "type": "string"
          },
          "ref_id": {
            "title": "Ref Id",
            "type": "string"
          },
          "snippet": {
            "default": "",
            "title": "Snippet",
            "type": "string"
          },
          "source_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Id"
          },
          "source_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SourceKind"
              },
              {
                "type": "null"
              }
            ]
          },
          "title": {
            "default": "",
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "index",
          "kind",
          "ref_id"
        ],
        "title": "ChatCitation",
        "type": "object"
      },
      "ChatContextItem": {
        "properties": {
          "concept_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Concept Id"
          },
          "content": {
            "default": "",
            "title": "Content",
            "type": "string"
          },
          "context_type": {
            "title": "Context Type",
            "type": "string"
          },
          "label": {
            "default": "",
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "context_type"
        ],
        "title": "ChatContextItem",
        "type": "object"
      },
      "ChatDocument": {
        "properties": {
          "chat_id": {
            "title": "Chat Id",
            "type": "string"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/ChatMessage"
            },
            "title": "Messages",
            "type": "array"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "session_id"
        ],
        "title": "ChatDocument",
        "type": "object"
      },
      "ChatMessage": {
        "properties": {
          "citations": {
            "items": {
              "$ref": "#/components/schemas/ChatCitation"
            },
            "title": "Citations",
            "type": "array"
          },
          "content": {
            "title": "Content",
            "type": "string"
          },
          "context_items": {
            "items": {
              "$ref": "#/components/schemas/ChatContextItem"
            },
            "title": "Context Items",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "message_id": {
            "title": "Message Id",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "role",
          "content"
        ],
        "title": "ChatMessage",
        "type": "object"
      },
      "ChatRequest": {
        "properties": {
          "context_items": {
            "items": {
              "$ref": "#/components/schemas/ChatContextItem"
            },
            "title": "Context Items",
            "type": "array"
          },
          "debug": {
            "default": false,
            "title": "Debug",
            "type": "boolean"
          },
          "message": {
            "maxLength": 8000,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "message"
        ],
        "title": "ChatRequest",
        "type": "object"
      },
      "ChatResponse": {
        "properties": {
          "assistant_message": {
            "$ref": "#/components/schemas/ChatMessage"
          },
          "chat": {
            "$ref": "#/components/schemas/ChatDocument"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/ChatCitation"
            },
            "title": "Citations",
            "type": "array"
          },
          "subgraph": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SubgraphResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "trace": {
            "items": {
              "$ref": "#/components/schemas/ChatTraceStep"
            },
            "title": "Trace",
            "type": "array"
          }
        },
        "required": [
          "chat",
          "assistant_message"
        ],
        "title": "ChatResponse",
        "type": "object"
      },
      "ChatTraceStep": {
        "description": "One structured step of the agent's execution (for the debug panel).",
        "properties": {
          "args": {
            "additionalProperties": true,
            "title": "Args",
            "type": "object"
          },
          "result_count": {
            "default": 0,
            "title": "Result Count",
            "type": "integer"
          },
          "step": {
            "title": "Step",
            "type": "integer"
          },
          "summary": {
            "default": "",
            "title": "Summary",
            "type": "string"
          },
          "tool": {
            "default": "",
            "title": "Tool",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "step",
          "type"
        ],
        "title": "ChatTraceStep",
        "type": "object"
      },
...
      "EdgeType": {
        "enum": [
          "MENTIONS",
          "RELATES_TO",
          "CO_OCCURS_WITH",
          "CONTAINS"
        ],
        "title": "EdgeType",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NodeType": {
        "enum": [
          "concept",
          "topic_cluster"
        ],
        "title": "NodeType",
        "type": "string"
      },
...
      "SourceKind": {
        "enum": [
          "pdf",
          "audio",
          "document",
          "image",
          "video"
        ],
        "title": "SourceKind",
        "type": "string"
      },
      "SubgraphEdge": {
        "properties": {
          "edge_type": {
            "$ref": "#/components/schemas/EdgeType"
          },
          "properties": {
            "additionalProperties": true,
            "title": "Properties",
            "type": "object"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "target": {
            "title": "Target",
            "type": "string"
          }
        },
        "required": [
          "source",
          "target",
          "edge_type"
        ],
        "title": "SubgraphEdge",
        "type": "object"
      },
      "SubgraphNode": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "node_type": {
            "$ref": "#/components/schemas/NodeType"
          }
        },
        "required": [
          "id",
          "label",
          "node_type"
        ],
        "title": "SubgraphNode",
        "type": "object"
      },
      "SubgraphResponse": {
        "properties": {
          "center_concept_id": {
            "title": "Center Concept Id",
            "type": "string"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/SubgraphEdge"
            },
            "title": "Edges",
            "type": "array"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/SubgraphNode"
            },
            "title": "Nodes",
            "type": "array"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "center_concept_id"
        ],
        "title": "SubgraphResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Corpus2Node API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/chat/message": {
      "post": {
        "operationId": "message_api_chat_message_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Message",
        "tags": [
          "chat"
        ]
      }
    },
    "/api/chat/stream": {
      "post": {
        "operationId": "stream_api_chat_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream",
        "tags": [
          "chat"
        ]
      }
    },
    "/api/chat/{session_id}": {
      "delete": {
        "operationId": "clear_history_api_chat__session_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatDocument"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Clear History",
        "tags": [
          "chat"
        ]
      },
      "get": {
        "operationId": "get_history_api_chat__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatDocument"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get History",
        "tags": [
          "chat"
        ]
      }
    },
...
    "/chat/message": {
      "post": {
        "operationId": "message_chat_message_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Message",
        "tags": [
          "chat"
        ]
      }
    },
    "/chat/stream": {
      "post": {
        "operationId": "stream_chat_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream",
        "tags": [
          "chat"
        ]
      }
    },
    "/chat/{session_id}": {
      "delete": {
        "operationId": "clear_history_chat__session_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatDocument"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Clear History",
        "tags": [
          "chat"
        ]
      },
      "get": {
        "operationId": "get_history_chat__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatDocument"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get History",
        "tags": [
          "chat"
        ]
      }
    },
...
  }
}